requirements.txt
docker>=6.1.2
paho-mqtt>=1.6.1
omegaconf>=2.3.0
pytest>=7.3.1
pytest-mqtt>=0.3.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
aiocron>=1.8
aiomqtt>=1.0.0